+++ mountpoint-s3-fs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Keep serving an implicit directory after Mountpoint deletes the last file in it, until it is removed with `rmdir`. ([#PRNUM](https://github.com/awslabs/mountpoint-s3/pull/PRNUM))
+
 ## v0.10.0 (July 20, 2026)
 
 * Add `tls_config` field on `s3::config::ClientConfig` so callers can configure a custom CA trust store through to the underlying S3 client. ([#1834](https://github.com/awslabs/mountpoint-s3/pull/1834))